chore: enable Biome noImportCycles rule
Set suspicious/noImportCycles to error so circular module imports fail
at lint time. Break the existing permission-tree <-> tree-node-header
cycle (and the wider cycle web through permission-card, util, and the
participants/add + pendingtasks pages) by extracting the exported types
(PermissionType, PermState, TreeNode, RefreshState) into a sibling
permission-tree-types.ts. Drive-by: move the misplaced biome-ignore-all
in nav-bar.tsx to the top of the file.

Refs #382
