+++ app/participants/[id]/page.tsx
@@ -8,7 +8,8 @@ import { useCsData } from '@/hooks/useCredentialSchemaData'
 import { usePermissions } from '@/hooks/usePermissions'
 import { useTrustRegistryData } from '@/hooks/useTrustRegistryData'
 import { useVeranaChain } from '@/hooks/useVeranaChain'
-import PermissionTree, { PermState, TreeNode } from '@/ui/common/permission-tree'
+import PermissionTree from '@/ui/common/permission-tree'
+import { PermState, TreeNode } from '@/ui/common/permission-tree-types'
 import { Role } from '@/ui/common/role-card'
 import { Permission } from '@/ui/dataview/datasections/perm'
 import { authorityPaticipants, nodeChildRoles, roleColorClass, roleJoinColorClass } from '@/util/util'
