+++ app/pendingtasks/page.tsx
@@ -5,7 +5,8 @@ import { faFolder } from '@fortawesome/free-solid-svg-icons'
 import { useEffect, useState } from 'react'
 import { useVeranaChain } from '@/hooks/useVeranaChain'
 import { usePendingTasksCtx } from '@/providers/api-rest-query-provider-context'
-import PermissionTree, { TreeNode } from '@/ui/common/permission-tree'
+import PermissionTree from '@/ui/common/permission-tree'
+import { TreeNode } from '@/ui/common/permission-tree-types'
 import { Permission, TrustRegistriesPermission } from '@/ui/dataview/datasections/perm'
 import { authorityPaticipants, roleColorClass } from '@/util/util'
